backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
reportlab==4.0.7
Pillow==10.1.0
qrcode==7.4.2
python-decouple==3.8
psycopg2-binary==2.9.9
